Add LeakyRelu

+++ projects/project2/project_road_segmentation/torch_unet/torch_unet/unet/train.py
@@ -90,7 +90,7 @@ def train_model(epochs, criterion, optimizer, lr_scheduler, net, train_loader, v
                     logger.info(f'Checkpoint {epoch + 1} saved !')
         
         if lr_scheduler is not None:
-            lr_scheduler.step(np.mean(epoch_loss) * 1000)
+            lr_scheduler.step(int(np.mean(epoch_loss) * 1000))
             writer.add_scalar("LR", get_lr(optimizer), global_step)
     writer.close()
     torch.save(net.state_dict(), os.path.join(dir_checkpoint, "final.pth"))
